Update Dominio Elementos, Historial, Tipo, Usuario

+++ Arquitectura_en_capas/CapaEntidad/HistorialElemento.cs
@@ -2,7 +2,7 @@
 
 public class HistorialElemento
 {
-    public long IdHistorialElemento { get; set; } 
+    public int IdHistorialElemento { get; set; } 
     public int IdElemento { get; set; }
     public int? IdCarrito { get; set; }
     public int idUsuario { get; set; }
